minimize register reads during firmware transfer
- Don't bother updating the 'get' variable if the last known value
  already allows for starting the next block.
- Don't check NCP status unless a timeout happens.

These two small changes result in a 76% reduction of the number of
register accesses per loop iteration (from 13 to 3.17) or a 70%
reduction of the total number of cmd53 sdio transfers.  This roughly
halves the time needed to transfer the firmware, which itself is the
dominant part of the wf200 initialization time.

Reduction in firmware download time measured on WGM160P:
- bare_metal example: 88 ms -> 48 ms
- micriumos example: 619 ms -> 207 ms